requirements.txt
dash==2.14.1
dash-bootstrap-components==1.5.0
plotly==5.18.0
pandas==2.1.3
Pillow==10.1.0
azure-cognitiveservices-vision-computervision==3.2.0
azure-core==1.29.5
python-dotenv==1.0.0 